Move Seal types.ts file
