+++ packages/widgets/src/widgets/SealModuleWidget/lib/types.ts
